enhancements for bug# 426940
The monitor script has been heavily modified to be multi-threaded.
A new table is introduced: 'priorityJobs'

The first thread (known as the standardJob thread) is the standard dump queuing
loop, it walks the file system seeking new dumps.  On finding a new dump, it
grabs the insertion lock, places the found job in the 'jobs' table, scheduling
it with a processor and then releases the insertion lock.  The standardJob
thread uses the lock to protect each insertion, so it will acquire and release
for each discovered new job.  Once it has finished the walk, it sleeps for two
minutes (configurable) before starting again.

The second thread is for priority jobs.  It scans the new table 'priorityJobs'
every five seconds (configurable).  This table consists of  just the uuids of priority
job requests.  Values are added to this table via a web interface.  When the
priorityJob thread finds values in the 'priorityJobs' table, it grabs the
insertion lock which stops the standardJob thread.   Then it checks to see if
any of the priority jobs are already in the queue.  It updates the priority of
any found in the queue.  For any jobs not found in the queue,  it starts its
own file system scan, seeking just the target priority jobs.  On finding them,
it queues them with a higher than default priority.  Once done, it releases the
insertion lock and goes back to monitoring the 'priorityJobs' table.



git-svn-id: https://socorro.googlecode.com/svn/trunk@366 efcc4bdc-162a-0410-a8af-5b4b46c5376d