Bug 417118 – Handle infinite loops and/or insanely large dumps gracefully Processor now edits some of the middle out of certain insanely large raw dumps before committing them to the database.  The edits are controlled by two new command line options: crashingThreadFrameThreshold (default 100) and crashingThreadTailFrameThreshold (default 10). When reading the raw dump output from breakpad_stackwalk, processor counts the frames that it sees of the crashing thread.  If that count exceeds the value of “ crashingThreadFrameThreshold” then it stops passing those frames on.  When the processor gets to the end of the crashing thread's frames, it goes back in the frame list by  “crashingThreadTailFrameThreshold” frames and passes those last frames on to the database.   The effect is to remove the middle, leaving the two ends intact.
Since the frames are sequentially numbered, the gap in the sequence serves as the indicator as 
to where the edited section was removed.

A new column has been added to the reports table: “truncated”.  This boolean just indicates if 
the aforementioned editing has occurred.

Example:
crashingThreadFrameThreshold = 7
crashingThreadTailFrameThreshold = 5
original crashing thread frames: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16
truncated crashing thread frames: 0, 1, 2, 3, 4, 5, 12, 13, 14, 15, 16



Bug 421940 – support passing cache path to minidump_stackwalk
The processor now has command line parameters that allow complete control over the invocation of 
the breakpad_stackwalk program.  “--stackwalkCommandLine” defaults to 
“$minidump_stackwalkPathname -m $dumpfilePathname $processorSymbolsPathnameList 2>/dev/null”  
Values of the substitution variables are inserted into the command line before the stackwalk 
command is invoked.  The substitution variables are simply other processor command line options 
or environment variables.  There are three valid forms for the substitution variables in the 
stackwalk command line: $variableName, $(variableName), %(variableName)s 

Example:
for caching version of minidump_stackwalk -
$minidump_stackwalkPathname -c /path/to/cache -m $dumpfilePathname $processorSymbolsPathnameList 2>/dev/null

it is acceptable to eschew all but the $dumpfilePathname subsitution variables:
/usr/local/bin/breakpad_stackwalk -c /path/to/cache -m $dumpfilePathname 
  /mnt/socorro/symbols/symbols_ffx /mnt/socorro/symbols/symbols_sea 
  /mnt/socorro/symbols/symbols_tbrd /mnt/socorro/symbols/symbols_sbrd 
  /mnt/socorro/symbols/symbols_os 2>/dev/null



Bug 421985 – when a report fails to be processed, give info as to why
The processor will now make an entry into the reports table even if processing that report 
failed. Four new columns have been added to the reports table: 'starteddatetime', 
'completeddatetime', 'success', 'message'.

'starteddatetime' – copied from the temporary entry in the 'jobs' table, this value 
                    indicates the time that processor started the job.
'completeddatetime' – copied from the temporary entry in the 'jobs' table, this value 
                      indicates the time that the processor completed the job.
'success' – a boolean indicating if the processor completed the job successfully or not
'message' – a message indicating any error conditions that occurred during the processing 
            of the dump.  This generally corresponds directly with an exception being thrown 
            in the Python program during processing.  The breakpad_stackwalk program does not 
            pass error messages on to the processor.  If a dump fails due to a problem 
            encountered by breakpad_stackwalk, this error is not reflected in this value.



Bug 411351 – should reprocess bad reports
Closely related to Bug 421985.  If a dump failed in processing then it is marked in the 
reports table as success == False.  Should this bug be resubmitted, the processor will 
attempt to reprocess it.  However, no facility has yet been created to resubmit a bug.  
It would have to manually be copied back into the monitor's dump filesystem queue.  
If processor encounters a dump that it has already processed and that dump's report has 
a value of success == True, then processor will decline to reprocess the dump.


git-svn-id: https://socorro.googlecode.com/svn/trunk@381 efcc4bdc-162a-0410-a8af-5b4b46c5376d