Added the creation of a symbolic link for each json file inserted into the filesystem. The symbolic link pathname is of the form:  {config.storageRoot}/index/{hostname}/{uuid}.symlink Any intervening directories that do not exist ("index" & {hostname}) will be created on demand.
This gives us set of indexes into the json/dump directories for a much faster access.  Monitor's 
standard allocation loop can look to these symbolic links rather than having to walk a tree 
searching for new file entries.  As the standard allocation loop finds a symbolic link, it 
fetches the actual file location and enters the relevant information into the database.  It 
then deletes the symbolic link.  The mere existance of a symbolic link means that it is a json 
file in need of processing. 
 
Finding the json file for a specific uuid that isn't already in the database (done by priority 
processing) now requires just looking through the various {hostname} directories for the 
appropriately named symbolic link.



git-svn-id: https://socorro.googlecode.com/svn/trunk@410 efcc4bdc-162a-0410-a8af-5b4b46c5376d