reworked to use the symbolic link indexing system instead of massive file system walks.
The collector now writes an index to the json/dump files.  They take the form of symbolic links save in a 
directory structure based on this template:  {storageRoot}/index/{hostname}/{uuid}.symlink  The hostname is
the name of the "webhead" that wrote the file.

The standard allocation loop now looks at listings in the {hostname} directory and uses the symbolic link to 
queue the actual json/dump file jobs.  There is no walking the tree looking for jobs.  Once a symbolic link has
been used to queue a job, the link is deleted.

The priority loop when given a priority uuid looks to the {hostname} directories to find its jobs.  Again, no
file system walking.

There is a new configuration file option: minimumSymlinkAge  This the the amount of time that a symbolic link
sits before it is eligible for processing.  Ideally to prevent the monitor from trying to process a symbolic
link before the collector has finished creating it, this value can be set quite low.



git-svn-id: https://socorro.googlecode.com/svn/trunk@413 efcc4bdc-162a-0410-a8af-5b4b46c5376d