In this exciting new version, the way that processors fetch jobs has been reworked to decrease the load on the database.  This also results in a change to the way that monitor queues priority jobs.  In addition, the processors' periodic registration with the monitor has been redone with a eye to fewer checkins. Finally, the granularity of the quit check has been increased for both processor and monitor.
Processor: controlled by the configuration parameter 'batchJobLimit' which defaults 
to 10K, the processor pull in many jobs in a batch from the 'jobs' table.  These jobs 
are then given out to the worker threads.

Each processor now has its own priority jobs table.  When the processor starts, it 
creates the table.  When processor quits, the table is dropped.  The monitor is 
responsible for populating the priority jobs table.  When it encounters a priority 
request, it places the target uuid into the assigned processor's priority table.  
The processor checks its priority table every 'checkForPriorityFrequency' seconds.  
When it finds a priority job, it makes sure that it is given to worker immediately.

Monitor: the monitor now places priority uuids directly into a processor's  priority 
jobs table.  It still uses the priority column in the 'jobs' table, but that is not 
the primary means of communicating the need for priority processing to the processors.  

When the monitor determines that a processor has died, it tries to drop that 
processor's priority table, just in case the processor didn't manage to do it 
before it quit.

New configuration option in .../processor/config.py:

checkForPriorityFrequency – the frequency in seconds for the processor to look at 
its priority table for new jobs

processorCheckInFrequency – the frequency in seconds for the processor to renew its 
registration with the monitor.  This number should be slightly smaller than the 
monitor's 'processorCheckInTime' configuration option to insure that the monitor 
does not inadvertently mark declare a live processor as being dead.

batchJobLimit – the number of jobs the processor should grab from the jobs table 
at a time.  The higher the number, the lower the load on the database.  However, 
as the number gets higher, the memory demands of the processor goes up.



git-svn-id: https://socorro.googlecode.com/svn/trunk@425 efcc4bdc-162a-0410-a8af-5b4b46c5376d