Bug 501148
written, awaiting staging in milestone 1.2

The previous version of Socorro had a configuration parameter called 'ignoreThrottleable' 
it has been renamed to the more proper 'neverDiscard'.

A new configuration parameter has been introduced called 'minimalVersionForUnderstandingRefusal'.  
It is a dictionary keyed by product name.  The value in the dictionary is the first version 
of the product that can fully understand and respond to the "Discarded" behavior.  For Firefox 
that value defaults to 3.5.4.  Products not appearing in that dictionary are assumed to not  
understand the discarding behavior.  For those products and for FF versions below 3.5.4, 
throttling will result in deferral rather than discarding.

Logging has also been improved.  In 'debug' logging level, individual actions for each submission 
are logged with uuid.  Of course, discarded submissions are not assigned a uuid. 


git-svn-id: https://socorro.googlecode.com/svn/trunk@1537 efcc4bdc-162a-0410-a8af-5b4b46c5376d