Bug 555729 - final backend change to complete the "lorentz" version changes.
Warning, nothing has been done to address changes in the database column length.
The column should be set to type 'text' with no length contraint, but the change
now would be very difficult.  Since that table will eventually go away with
the new hbase/hadoop stuff, it's not worth the effort.


git-svn-id: https://socorro.googlecode.com/svn/trunk@1929 efcc4bdc-162a-0410-a8af-5b4b46c5376d