added a more sophisticated submitter.py.  This one will eventually replace the one in .../socorro/collector.  For now, I'm leaving that one in place.
  New features:
        -s, --delay
                pause between submission queing in milliseconds (default: 1)
        -D, --dryrun
                don't actually submit, just print product/version
        -N, --numberOfSubmissions
                the number items to submit (all, forever, 1...) (default: all)
        -n, --numberOfThreads
                the number of threads (default: 4)

the --numberOfSubmissions option allows the submitter to get
a specific number of crashes from the source (specified by --search or -s), or
loop through a search over and over.  The default behavior is to
just submit all the crashes found in a search (this mimics the behavior of the
old submitter).

This app is BETA.  There are known defects:
  * the --numberOfSubmissions option: all and forever work fine, integers do not.
  * khan seems to have a defective pycurl module for python26.  Actual
    submissions are not tested.



git-svn-id: https://socorro.googlecode.com/svn/trunk@2841 efcc4bdc-162a-0410-a8af-5b4b46c5376d