Bug 641909 - added a new processor registration mode called 'host'.  The current system allows multiple processors to contend for taking over the same dead processor.  By setting the configuration parameter 'processorId' to 'host', the new processor will only look for a dead processor that came from the same host as the new processor.
If a dead processor is found with a match host name, the new processor
will take over for the dead one.

If a live processor is found for the host, the new processor will refuse 
to run.  

If no processor is found at all for the host, the new processor will 
register as a new processor.


git-svn-id: https://socorro.googlecode.com/svn/trunk@3475 efcc4bdc-162a-0410-a8af-5b4b46c5376d