Updated the crashMover
After enabling exception logging in the os.walk method, a lot of lines
about absence of the self.root/self.dateName folder was created in the
log file.
This was happening because of a socorro_fs.cleanEmptySubdirectories() call
with top_level parameter set to self.root.
I think it's better to set it to self.root/self.dateName