Add github to vcsMappings
I'm going to be fixing the symbol files for B2G soon so that they contain the right repository info. To make this useful we'll need to be able to translate these into web links. As a first cut, making github links work will be useful. (Submitted this same patch for the php frontend at https://github.com/mozilla/socorro/pull/891 .)