rollback() unconditionally in the transaction executor.
If we're not in a transaction, nothing bad's going to happen. PG gives a NOTICE in the case of rollbacks and a WARNING in the case of commits, and we've been doing commits unconditionally forever on master.

Delete in_transaction(), which becomes unused.

Delete a test that was just making sure we didn't roll back when in a transaction. I don't think it asserts anything else useful (a docstring would have been nice here); test_rollback_transaction_exceptions_with_postgres() already makes sure we roll back on exception.

Update a test's expected rollback count; rollback always happens now, harmlessly.