(no bug) slightly improved crontabber-state layout
The crontabber-state layout now adjusts to the size of the window at the time the page is opened instead of being a static 980px.

The individual jobs are now positioned in such a way that jobs 'float' as far left as they can, which results in a more even layout for the typical crontabber dependency graph. No more pyramid shape.