EOD temp save. getting closer to tests file being finished
This gets most of the set up complete for non-webapp tests. The webapp
lines are little bombshells -- they need to be reworked because some
duplicate work there is in the top 3 time sinks for the build process.
It shouldn't be nearly as involved as this but still more work than
two lines implies.

Lots of good stuff in here, though. Syntax has improved, UX for devs
is far better, and everything is pretty linear. It should be much more
maintainable. I'm moving an exit 0 along the process so I can eisily
see where I am at in the process. Shouldn't be long now.