deploy: Ensure that cron file doesn't have +x.
crond on my CentOS 6 box refuses to load files with the executable bit.
Since chmod-ing does not update the mtime (and cause a crond reload) of
the file, we chmod it before we cp.

Not sure where the +x bit came from - it's not in git - but deploy.sh
installed one with an executable bit on my system.