Footer overhaul part one
This patch reorganizes footer links to remove redundant links and to colocate
all links that require authentication. Also makes the footer text more concise,
and swaps the dull dino logo for a 2-tone cutout version. Removes the fixed
height and adds a float clearing break after the content to ensure the footer
box contains all of the floated elements, even on narrow pages where they would
previously spill out.

Makes some space and formatting fixes to the base template so its easier to
match up opening and closing tags visually. Also converts tabs to spaces in the
base template

Fails to pin the footer to the bottom of the page when there isn't enough
content. That will have to wait for part two.