Add support to stackwalker for fetching symbols via HTTP
This changeset adds two new commandline arguments to the stackwalker:
  --symbols-url: to specify a base URL for fetching symbol files
  --symbols-cache: to specify a directory in which to save fetched symbol files

If either argument is provided they must both be provided. If neither is
provided the stackwalker will simply use any local paths listed after the
minidump argument as it currently does.

If both a symbols URL and local paths are provided the stackwalker will search
for symbols in the following order:
1) In the symbols cache directory
2) In each local path provided in order
3) By constructing a URL from the symbols URL and the symbol filename and
   GETting it, storing it under the symbols cache directory.