move search_util into the middleware module
search_util is consumed only by the middleware and doesn't really need to be in the common_lib at all. This sets a pattern for other common bits, notably those in `socorro.external` -- elevate common code to the least-generic module. If it is consumed only by the processor, move it to the processor. If it is used by processor and collector, move it so socorrolib.

Because these changes are roughly hewn, we're probably going to find more of these modules as the integration gets further along. That's alright -- like pep8 this is a guideline to follow as we notice problems and not necessarily a stop-the-world-to-garbage-collect change.