Fixes bug 1266827: Add keybuilder class to s3 crash storage
This breaks out key building into a separate class allowing us to
specify which keybuilder class to use.

It creates a new DatePrefixKeyBuilder that takes the date portion from
the end of the crash_id and an entropy portion from the beginning of the
crash_id and puts those two things before the crash_id in the
pseudo-filename.

This new pseudo-filename has entropy to reduce problems with s3
partitioning, but allows us to generate a listing of all raw crashes
submitted on a given date (takes 4096 bucket list requests, but it's
doable).

This is backwards-compatible with the current pseudo-filenames because
it changes .build_key() to .build_keys() which returns a list of keys to
try in order of "goodness". That way .fetch() can try the keys in order
until it finds the object in question.