Bug 1292235 update socorrolib to 0.2.3
I reduced the requirements.txt file for socorrolib to just the
requirements that it needs. The requirements.txt file is sourced as
install dependencies by setup.py. Because of this, it'll be tricky to
update dependencies that socorrolib doesn't actually use in this
repository.

This updates to the latest version of socorrolib.