Bug 1293776 initial pass on StatsdMetrics (#3439)
* Bug 1293776 initial pass on StatsdMetrics

This creates a new kind of external component called Metrics. It gets
used in the breakpad collector after saving the crash to log data
about the incoming crash at the point we've collected it.

* Bug 1293776 adjust crash size report metrics

This adjusts the crash size report metrics such that they also capture
data about rejected crash reports. Further, it tweaks the data key so
it's easier to intuit what it is in datadog and other similar systems.

* Bug 1293776 Fix integration test

I think this fixes the integration test that Travis runs using the
Collector2015App that we don't use. That app re-arranges where
configuration is located in the collector app hierarchy of components,
so "config.metrics" gets handled differently.

* Bug 1293776 address review comments

* fix docstring
* move the rest of the metrics code into the try/except block
* reduce the work the except block does so as to reduce the likelihood of errors
  in the error handling