force body to take up the screen
critical to the sticky footer is forcing the body element to take up the whole height of the window. Typically this is accomplished by setting height:100% on the html element, but with our header setup this results in the header being crushed to it's min-height (0) on search and possibly other pages.

Instead I'm using vh units to force the body size to 100%. This is clearer in intent and has been supported in FF since 19.