fixed bug 1290967 (#3454)
Some linux crashes are crashing in RunWatchdog without calling it from or otherwise including the '::' operator in the frame, so these crashes are not triggering the appropriate processor rule.

This fixes it by removing the '::' requirement from the predicate. There are no Firefox crashes in super search at the time of this writing that will be adversely caught by this -- it should catch the same crashes as the previous rule plus the missing reports.