remove the ServerStatus model from alembic
The class had a typo in the name, but I don't believe the class name was important. It was introduced as part of the schema.sql replacement and modified twice since then, once in 2013 and once in 2014.

From the other modifications I don't believe this change requires a migration. Additionally, there are is no equivalent test file to modify. The class in tests that checks the output of this model is already modified to remove processor status in e77f82a44ccb9af8ebe5903a142ebf5e60977cc5