fixes bug 1301222 with a transform rule (#3462)
* fixes bug 1301222 with a transform rule

add a transform rule to mutate the default theme's extension id into a more easily identifiable string

* (no bug) - moar docstring for ThemePrettyNameRule

Add an additional caveat to the docstring of the ThemePrettyNameRule

* bug 130122 - fixup for reviewer comments

alters the prettified name of the default theme according to nnethercote's PR comments over the suggestion in the original bug

* bug 1301222 - reviewers comments 1

rework the ThemePrettyNameRule to use a dictionary of conversions
rather than detecting one specific conversion.

* (nobug) - pep8 mozilla_transform_rules

pep8 fixes to make mozilla_transform_rules comply with the style guide

* Bug 1301222 - Added tests for predicate.

* bug 1301222 - reviewers comments 2

This alters the predicate behavior in the case that a processed
crash is missing an addons property to use a null list rather than
test for the property. This preserves the structural similarity
between the predicate and action methods without sacrificing the
safety of the test method.