1300367 remove server status (#3456)
* remove server status api docs

The server status api endpoint was removed in e77f82a44ccb9af8ebe5903a142ebf5e60977cc5 and this removes the documentation for the now non-existant service.

* remove the database schema docs

The database schema pdf was last generated nearly four years ago and the graphml equivalent has gone more than 18 months without update. They are growing increasingly incorrect and this patch removes them.

They can be recovered from history, if needed.

* remove server_status from development docs

the server_status table is deprecated and should be removed from the developer documentation

* remove server_status crons

remove the server status crons and references to them in configuration. oddly there were both crontabber and non-crontabber versions of the cron lying around. The one without tests predates our adoption of rabbit and has probably been unused quite a while.

* remove the ServerStatus model from alembic

The class had a typo in the name, but I don't believe the class name was important. It was introduced as part of the schema.sql replacement and modified twice since then, once in 2013 and once in 2014.

From the other modifications I don't believe this change requires a migration. Additionally, there are is no equivalent test file to modify. The class in tests that checks the output of this model is already modified to remove processor status in e77f82a44ccb9af8ebe5903a142ebf5e60977cc5

* remove the current_server_status_view

Removes the server_status view from the raw_sql directory. This probably needs a migration to accompany it.

* migration to remove server_status table

this removes the server_status table from models.py and provides
a reversible alembic migration for our instance

* remove server_status from the endpoint tests

while the endpoint for server_status is still around, it has not
relied on the table to operate since e77f82a. This removes the step
from the server_status test teardown that touched the unused
server_status table