No Bug - Hide loader only when loading is done.
QA tests rely on the loader gif element to disappear to start running their tests. As of this, that loader will get removed only when the dynamic_form library is done with everything it has to do. That will prevent weird race conditions to happen, and hopefully will make the tests more reliable.