Fixes bug 1297760: handle OSError when symlinking
We have a problem with the -stage collector where we can get two
crashes with the same crash id. The first one will come in and get
saved--no problem.  The second one comes in and then trouble
starts. First, save_raw_crash() saves the crash by calling into
_save_files(). _save_files() happily stomps on any existing file, so
if there was a file already, it's no big deal.

Then save_raw_crash() goes to create two symlinks. Before this patch,
it would raise an OSError creating the name symlink because it created
one already with the first crash. That wouldn't get handled, so it'd
percolate up all the way and cause the collector to return an HTTP 500
to the client. Blech.

This patch changes the code to ignore the OSError. The better thing to
do would be for FSTemporaryStorage to realize the crash already exists
and stop trying to save it again, but that would require non-trivial
refactoring of the code. Instead, what this does is put both symlink
creating method calls into a try/except to ignore the OSError that
would result if we try to create the same symlink a second time.

A while back, someone had a similar error and wrote a test to handle
the case where one symlink is created, but not the other one. This
nixes that and replaces it with a test for making sure
save_raw_crash() doesn't raise an OSError due to symlinks.