fixes bug 1311697 - deep copy SocorroDotDict better (#3546)
* fixes bug 1311697 - deep copy SocorroDotDict better

* Fix dictify

This reworks dictify to more obviously traverse an object graph
converting mapping-ish things (including DotDict and DotDict) to dicts
and sequence-ish things to lists.

The guarantee after the conversion is that the resulting thing is
deep-copyable.

This also adds a bunch of tests for the various possible cases.

* Fix a gross error

I am deeply sorry such an egregious error made it into the previous
commit. I hope future Socorro developers can find it in their hearts to
forgive me.