Fixes bug 1305129: GA ssl handshake fix (#3560)
* Fixes bug 1305129: GA ssl handshake fix

Certificate verification is failing. We debated several options (details
in the bug) and then decided to set verify_ssl=False for this specific
call. Because we're doing that, we nixed all possible PII, too.

* Fix tests

This fixes the tests and also removes the client_id argument.

* Rewrite uuid mocking in tests

This cleans up the uuid mocking by converting the repeated stuff into a
decorator.