Fixes bug 1318448: reduces deepcopies we do in the processor (#3594)
* Fixes bug 1318448: reduces deepcopies we do in the processor

This changes the PolyCrashStorage code such that it only does a deep
copy of the processed crash if the crash storage says it mutates the
crash data.

Currently, we have 5 crash storage classes configured in production.
This change will reduce the number of deepcopies we're doing from 5 to
1.

* Fix for statsd wrapped crash stores

* Fix the fix

The previous fix didn't do the right thing. This does.