Fix is_mutator() call (#3596)
The StatsdCounter is a duck-typed crash storage so it doesn't have an
is_mutator().

So what we're saying now is "if the crash storage has an is_mutator and
that returns True, then deepcopy". The underlying assumption is that
crash storages without an is_mutator aren't mutating the crash data.