Fixes bug 1319896 - Increased the number of shards per index to 10. (#3632)
* ES_SHARDS_PER_INDEX => 10 (was 5)

* fix test_crashes_per_day_failing_shards 40% => 20% with new ES_SHARDS_PER_INDEX

* increase shards per index to 10 from default 5

* Fixes bug 1319896 - Increased the number of shards per index to 10.

Refactored the ES settings creation to be in the index_creator file instead of super_search_fields where it did not belong. Put the number of shards into a configuration option, which then required that the IndexCreator class was
imported via configman, or the option would be unknown, causing errors.