Fixes bug 1331726 - fix assertion in test_public_api_navigation (#3660)
This fixes the assertion that we should have at least one signature first date
for each signature--that turns out to not necessarily be the case. I changed it
to make sure it's > 0. Hopefully we don't eventually find ourselves in a
situation where it's 0, though I suspect that's possible.

While doing that, I changed some variable names to make the code easier to read
and I added comments to summarize what's about to happen so it's easier to
verify the test is doing what it should be doing.