Fixes bug 1363130 - remove Collector2015App, BreakpadCollector2015
The Socorro collector had a 2015 version of a few classes that had a different
url registration and allowed for collecting of other non-Breakpad things. The
classes were used in the integration tests and also had unit tests.

We (Mozilla) never ended up switching to that, so this removes all of it because
we're not going to use it and it's no longer used in the integration tests.