Fix postgresql tests (#3769)
* Fix postgresql tests

The postgresql server_status test stomped would overwrite two files, run the
test, and then delete the files altogether. This fixes that behavior by mocking
out the "read the file from disk" function so we're still testing the code
itself, but not stomping on the files in question.

There was another postgresql test that had the database host, username, and
password hardcoded. This changes those to use the relevant environment
variables.

* Remove unused tempfile