Fix logging in startup so it tell us errors
At startup, there is no self.config.logger, so if things go awry, then calling
self.config.logger.critical doesn't work and then you only see that problem and
not the original problem that occurred which is pretty unhelpful.

Is this a great fix? No. But it's better than before.