Rework tests to run in circle ci
Circle supports docker-compose, but you have to do some things to get it
working. One thing it doesn't support is volumes, so you have to create a data
container, copy stuff into it, and then use that.

This implements a run_tests_with_docker.sh script that builds up the docker
environment for running the tests that'll run both in circle ci and local
development environment.