Bug 1366008: Fix circle tests (#3790)
* circleci conf: test building docker images in circleci

* circleci conf: add a circleci v2 config.yml, deprecate v1

* circleci conf: v2: use remote docker in circleci, switch to ubuntu base container, install docker manually

* circleci conf: add working_directory

* circleci conf: fix localstack and postgres image locations

* circleci conf: install misc dependencies in base image

* circleci conf: install docker-compose

* circleci conf: using working_directory /socorro

* circleci conf: temporarily skip dockersetup/dockertest

* circleci conf fix perms on bin/ci/deploy-dockerhub.sh

* Remove old circle config file and re-add test step

* circleci conf: s/build/latest, +x in bin/ci/deploy-dockerhub.sh

* Rework tests to run in circle ci

Circle supports docker-compose, but you have to do some things to get it
working. One thing it doesn't support is volumes, so you have to create a data
container, copy stuff into it, and then use that.

This implements a run_tests_with_docker.sh script that builds up the docker
environment for running the tests that'll run both in circle ci and local
development environment.

* Fix pushing to dockerhub

We don't have a socorro_test image to push anymore.