Bug 1361764 - initial switch to pytest (#3787)
* Bug 1361764 - initial switch to pytest

This switches the test runner from nose to pytest for socorro/unittest/
tests (i.e. not the webapp tests).

This doesn't get rid of nose completely--we still have to go through and remove
ok_ and eq_ instances. You can continue to use nose if you like.

This fixes tests that were broken in pytest so that they work with both pytest
and nose.

* Fix env vars for local running

* Fixed tests when ran locally, without any env variables.

* Now using pytest-env instead.