Fixes bug 1368698 - Improve tests infrastructure for dev (#3803)
This improves the test infrastructure in docker for development. It adds a
"dockertestshell" make rule which sets up the docker environment with the repo
directory mounted in the container and then starts a shell. With this,
developers can run individual tests and also pass in different arguments to the
test runners.

This also documents all that in the docker README.