Fixes bug 1368498 - Do not use bugs table in bugzilla cron job. (#3804)
* Fixes bug 1368498 - Do not use bugs table in bugzilla cron job.
This makes the bugzilla cron job stop writing any data into the bugs table. It also removes the foreign key from bug_associations to bugs in a migration, which is needed for the cron job to keep working.
It also removes the call to the ``reports`` table, making the job write all associations without verifying if the signatures exist in our system. This happened because 1. we want to get rid of the reports table and 2. I believe this was a source of bugs (with signatures incorrectly missing in that table).

* Some cosmetic changes, and removed useless tests.

* Removed unneeded data from bugzilla.

* Made tests more robust.